cairo-corpus-starter/requirements.txt
beautifulsoup4
requests
tqdm
python-dotenv
orjson
rapidfuzz
jsonschema
pytest
